backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asyncpg==0.30.0
click==8.3.1
colorama==0.4.6
cryptography==45.0.0
email-validator==2.3.0
fastapi==0.128.0
h11==0.16.0
httpx==0.27.0
idna==3.11
passlib==1.7.4
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.0.1
pytest==9.0.2
PyJWT==2.10.1
starlette==0.50.0
typing_extensions==4.15.0
uvicorn==0.40.0
